Introduce error handling in user and video addition
Added error handling in the process of user and video addition to the database. This update will ensure that in case of a failure in adding users or videos to the database, an error message is printed. The function will then exit or continue, preventing potential issues down the line.