feat:Add functionality to update user avatar in database
Implemented a new function in db.rs to add or update the avatar URL for a YouTube user in the database. This function is then called in rss_youtube_api.rs whenever a successful YouTube fetcher request for a user avatar URL returns, thereby keeping the database up-to-date with the latest avatar URLs.